finished JSONread
Can be executed by calling:
readBattIO() - shows the in/out of batteries in amps
readBattpercentage() - shows the battery percentage
readSolar() - gets Watts of solar energy.

All returned as arrays
