--- PowerHouse/JSON read/.vscode/settings.json
@@ -1,6 +1,0 @@
-{
-    "editor.tokenColorCustomizations": {
-        "comments": "",
-        "textMateRules": []
-    }
-}
