--- PowerHouse/JSON read/testfile-single.json
@@ -1,1 +1,0 @@
-{"time": [2023, 10, 10, 13, 28, 21, 1, 283, 1], "topic": "solar/stats", "message": {"function": "READ", "model": "RNG-CTRL-ROVER", "battery_percentage": 100, "battery_voltage": 13.4, "battery_current": 13.77, "battery_temperature": 22, "controller_temperature": 36, "load_status": "off", "load_voltage": 0.0, "load_current": 0.0, "load_power": 0, "pv_voltage": 71.5, "pv_current": 2.68, "pv_power": 192, "max_charging_power_today": 371, "max_discharging_power_today": 0, "charging_amp_hours_today": 41, "discharging_amp_hours_today": 0, "power_generation_today": 550, "power_consumption_today": 0, "power_generation_total": 84472, "charging_status": "mppt", "battery_type": "lithium"}}
